pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "adafruit-circuitpython-neopixel>=6.3.12",
    "board>=1.0",
    "pygame>=2.6.1",
]
